RSDEV-701 Add About RSpace dialog and page (#438)
## Description ##
This change adds a new "About RSpace" section that is both added to the
account menu in the app bar as a dialog and in its own standalone
"/about" page (which isn't linked anywhere but may at times be helpful
for providing support).

## Testing notes
Either go to "/about", or click on the "About RSpace" menu item in the
app bar's account menu.