RSDEV-734 Add skip-to-content menu to Gallery and Inventory (#439)
## Description ##
For user who rely on keyboard navigation, having a menu that
automatically appears when they press tab to skip to particular parts of
the page is a major improvement to the usability of the page as they no
longer have to press tab over and over to move passed structural content
such as the app bar, account menu, help, etc.

## Testing notes
On either the Gallery or Inventory, press "tab" to see the menu, arrow
keys to select a landmark, and then "enter" to select it. Thereafter,
the tab focus will be moved to the landmark, from which continuing to
press "tab" will move the focus into the landmark's content.