old gallery snippet sharing
The share dialog does not show the current shared status of the snippet,
and thus there is no way to unshare, change the permission, or change
the folder in which the snippet has been shared. This is due to the fact
that the /api/v1/share GET endpoint does not seem to allow snippet
global ids in the query parameter. This is not worth fixing as we need
a new endpoint that specifically gets all of the shares associated with a
particular record.