PRT-979 Fix callToAction colour in high contrast mode (#456)
## Description ##
This changes fix the issue where the buttons with the `callToAction`
colour would not have sufficient contrast when the user has enabled a
high contrast mode in their browser.

In addition to fixing the issue, this change also adds a test for
ValidatingSubmitButton, one of the main use cases for the `callToAction`
colour that asserts that the button has sufficient contrast to meet the
AAA threshold of the WCAG 2.2 standard when the browser is requesting
additional contrast. This is not simple to do because the MUI Button's
ripple effect normally prevents automated accessibility testing tools
like axe from being able to assert the foreground and background colours
as they cannot be certain what text is actually visible. As such, we
remove the ripple to be able to assert the colours.