PRT-985 Fix calls to trackEvent when not available
Because we're using the $(this).click event handler, and global.js is loaded
inside of the internal link picker dialog, this code was executing inside of
the dialog, where trackEvent is not available. jQuery always returns a truethy
value even if it finds no elements that match the selector so the contional
check wasn't even doing anything; a pointless check since the day it was added
almost 7 years ago.

Initially, I tried to remove the use of $(this), which refers to window, but
that would have meant finding an alternative, robust point in time to attach an
event listener to the HTMLabelElement that is the parent of the checkbox. I
couldn't find a good way to do that, so I just left the $(this).click in place.

So instead, now, we're still firing an event every time the user taps anywhere
in the window, but if the are no .displayRevisions -- as there aren't in the
internal link picker dialog -- then we just don't call trackEvent. This, in
itself, uncovered another bug however as we were firing the analytics event
everytime the user tapped anywhere in the window, even if they hadn't changed
the value of the .displayRevisions checkbox. So now we also check to see if the
target of the click event is the checkbox itself before calling trackEvent.

An alternative approach, as I applied in the other part of this change, is to
simply gate calls to trackEvent on trackEvent being defined. However, I did not
want to suggest that trackEvent is not available on the main document editor,
where the .displayRevisions checkbox is located, by inserting such checks into
the logic that should only be executing on the document editor. Instead, we can
now be sure that the code logging events about showing the last modified
information below each document field only runs in the context of the document
editor.