Rsdev 689 stoichiometry backend (#457)
## Description ##
This PR contains the backend changes for Stoichiometry including CRUD
operations, DB migration to create the new tables, code to copy
Stoichiometry as part of a document copy, and changes to get additional
info from the new `chemistry` service endpoint.

Requires the latest version of the `chemistry` service to run.

## Testing notes
The branch `rsdev-689-stoichiometry-ui` contains these back-end changes
plus the UI which could be used for testing purposes. We are holding
back on the UI PR until the UI branch has been further tested by
chemists.

Otherwise, swagger specs have been added for the Stoichiometry endpoint.
To use them, create a document and add a chemical element. Check the DB
or network tab to get the `RsChemElement.id` and use that as the
`chemElementId` when creating a Stoichiometry.


related `rspace-core-model` PR:
https://github.com/rspace-os/rspace-core-model/pull/33

---------

Co-authored-by: rs-fraser <fraser.macinnes>