PRT-989 Fix layout of Grid container dimension field in Firefox
For some reason Chrome and Firefox have a different behaviour when a flex box
with columnar layout has wrap enabled and has content that can be laid out in
multiple columnss. In Chrome the content is laid out in a single column, while
in Firefox it is laid out in multiple columns. This commit forces a single
column layout in Firefox by setting flex-wrap to nowrap.