RSDEV-826 Lazy load KetcherDialog and GalleryPicker (#482)
## Description ##
Ketcher is a very big third-party library that dwarfs all of our other
libraries and our own code. It is also used in several places across the
codebase so was being included in several JS assets. Moreover, it is
only used by a minority of users, and yet all users were paying the
performance cost of downloading and parsing this library's code several
times over.

This change instead lazily loads the KetcherDialog component, only
downloading the JS code when the user goes to use the component. Webpack
will now create a separate JS asset for it, which should be cached by
the users' browser meaning that they will only need to download the code
once across our entire product.

The only downside being that there will be a brief spinner the first
time that the user goes to use the component, but this is certainly
preferable to adding the same loading to the initial page load for
everyone. If this is intolerable, we could look at pre-emptively loading
the JS asset if the user has the chemistry integration enabled (although
this would mean gating the Inventory chemistry previewer behind the
chemistry integration which currently it is not).

I also added lazy loading to the usage of the Gallery in the
Stoichiometry dialog and to reduce its bundle size too.