Fix bug where TinyMCE had only one toolbar (#481)
## Description ##
This change fixes #475 wherein it was possible for the user to remove
the second TinyMCE toolbar, which the code that adds toolbar buttons for
the various integrations assumes will always be there. This PR makes two
changes
1. When the user is customising the toolbar, they are prevented from
removing the second toolbar
2. When registering toolbar buttons for the various integrations, we now
create additional toolbars to ensure that we can add the new toolbar
buttons to the second toolbar. This catches cases where users have
already broken their toolbars or where the second toolbar may be removed
through any other mechanism.