RSDEV-828 Use HeadingContext to dynamically adjust TitledBox heading level (#485)
## Description ##
Success criteria 2.4.10 of WCAG 2.2 requires that headings are organised
hierarchical, such that a h3 only comes after a h2. This allows screen
reader users to efficiently navigate by headings. TitledBox is a custom
react component used for organising some pages of the product.

TitledBox's heading level should dynamically adapt to be the next level
down from whatever the parent heading level is, wherever it is rendered.