Code cleanup, minimising dependency cruise warnings (#483)
## Description ##
We use dependency cruiser to enforce some structure to the frontend
codebase. Errors are surfaced when dependency cruiser is run in Jenkins
but warnings are not, so they've piled up somewhat due to the recent
changes to the codebase. This change re-organises some files, updates
the dependency cruiser config, and marks an old file as deprecated to
minimise the number of dependency cruiser warnings.

I suggest reviewing the changes commit-by-commit as each change is
unrelated from the others.

The command from running dependency cruiser is in package.json