PRT-996 Fix wrapping of long dd elements in DescriptionList component
"1fr 1fr" means that the container will have two columns of equal width, with
Firefox correctly implements. Chrome, on the other hand, has a bug where it
implements the slightly different behaviour of shrinking the width of the
columns when one of them has more content than fills 50% of the container
width. This means that if one column has a lot of content, it can cause the
other column to shrink to fit the remaining space, which is actually exactly
what we want to minimise text wrapping. To get this behaviour in both browsers,
I simply changed the first "1fr" to "auto", thereby shrinking the first column
to fit its content, and allowing the second column to take up the remaining
space.