RSDEV-821 Automatically appending asterisk to inventory searches (#513)
## Description ##
Query strings used in searches in Inventory are exact matches for names
of containers, samples, etc. To search for an item based on a prefix
match, an asterisk has to be appended to the query strings, but this is
not very discoverable. This change automatically appends an asterisk
when there are no search results.