RSDEV-464 fix grid labels being ignored in list view (#508)
## Description ##
The Grid Coordinates being displayed in a container's list view did not
respect the container's "Grid Labels" settings, and instead displayed
the Row and Column in a fixed format. This PR fixes this behavior and
the Grid Labels configuration will now be respected.

## Testing notes
1. In the Inventory system, create a new Container with type "Grid".
Note the Grid Labels have been pre-set to Row: ABC and Column: 123.
2. Create a subsample in the container.
3. Switch back to the container's view, under "Locations and Content",
switch to the List view.
4. Under List view, add the "Grid Coordinates" column to the view.
5. The Grid Coordinate should now display the coordinate instead of "Row
X of Y, Column X of Y"