RSDEV-762 Redesigned share dialog (#500)
## Description ##
This new share dialog allows users to view the current shared status of
the selected document, as well as more easily share one or more
documents with their groups and collaborators.

## Backend changes
On the back-end, there are a few new API endpoints to support the new
share dialog:

- move document
- the code for performing a move and all of the required checks was
contained within the WorkspaceController, so I created a
WorkspaceService and moved only the move document code to the new
service and adapted it slightly. Eventually we will be migrating all
workspace functionality for reactification so I made minimal changes
since we will be changing it all in the fairly short-term
- get all shares for document
- update permissions for an existing share
- get group by id

There were some other changes:

- remove `@Min(1)` validation for group and user share requests, so they
work with default users
- removed `UserSessionTracker` parameter that was passed down various
methods but unused

`rspace-core-model` PR:
https://github.com/rspace-os/rspace-core-model/pull/43

---------

Co-authored-by: rs-fraser <141634676+rs-fraser@users.noreply.github.com>