RSDEV-773 Support previewing snippets in Gallery (#505)
## Description ##
When the user is browsing the Gallery, it should be possible to view the
contents of the snippets. This is functionality that was never migrated
over from the old gallery to the new and is a blocker to support snippet
sharing and migrating the gallery picker user in the document editor.