RSDEV-428 Change document editors's Gallery picker to use new Gallery (#502)
## Description ##
The file picker for choosing Gallery files when editing a rich text
field in the ELN's document editor now uses the new Gallery, rather than
the old, removing a blocker on removing the old gallery.

Note that whilst this change creates a new TinyMCE plugin to support
this new picker, it utilises the existing jQuery code to generate the
required HTML to insert into the document field based on the file type.