RSDEV-756 show deployment-defined help links (#519)
## Description ##
Shows a list of help links provided by the institution, if defined in
deployment config.

Note: There's a bug where opening the menu in the Workspaces section
would cause the menu to open at the bottom of the page. As other styling
in the App Bar are also not present in the page, the bug is likely not
caused by this PR.