Rsdev 860 snippet path to root (#520)
## Description ##
Fix bug where parent folder and path to root information for a folder
wasn't being populated unless the user was the owner of the parent
folder.

Instead, check if the user is the owner or if they have access in a
shared context.

Handles the case of a notebook, which could have multiple contexts for
the same user if it's in their workspace and in a shared folder by
optionally accepting a `parentId` to distinguish.

https://researchspace.atlassian.net/browse/RSDEV-860