Rsdev 878 snippet api (#537)
https://researchspace.atlassian.net/browse/RSDEV-878

## Description ##
This PR adds 2 new GET snippets endpoints - 1 for retrieving the snippet
and associated metadata, and 1 for retrieving only the content.

For the snippet sharing work, which is the last blocker to fully
removing the old gallery, we need to be able reference snippets in
folder contents via the common `link` fields, which enables the
front-end to retrieve the snippet. The current snippet endpoints are of
the old style, non-public endpoints hence the need to add public REST
API endpoints.

Also updated the 1 usage in react code of the old style endpoint to use
the new REST API endpoint. And marked the old `SnippetController` as
deprecated. The rest of these methods will be migrated as creation etc.
gets moved once the document editor is react-ified. In
`SnippetController` I left the original `/content` endpoint in place, as
it's still used by the old gallery, but can be removed once the old
gallery is.

## Design decisions
We might not need both endpoints (could use `content` field from the
other endpoint), but I decided to keep the `/content` one for
convenience.

## Testing notes
The 2 new endpoints can be tested directly via swagger.

The update for the front-end to use the new `/content` endpoint can be
triggered via (new) Gallery → Snippets → double click snippet to
preview.