Rsdev 296 snippet sharing (#544)
https://researchspace.atlassian.net/browse/RSDEV-296

## Description ##
This PR adds the required changes to allow snippets to be shared via the
new react gallery.

1. `/groups/{id}` endpoint includes `sharedSnippetFolderId`
2. `/share/document/{id}` returns sharing information for snippets
3. `/api/v1/folders/tree/{id}` list snippets and works with gallery
shared snippet folders

I also tidied up setting parent ids in the response from 3. Previously,
the `parentId` of the folder being listed was only set when there was
only 1 parent, which isn't the case for shared snippet folders. Now we
use the new `FolderNavigationService` which handles edge cases. Then for
the child items, `parentId` is just set as the folder the tree request
is made for.

## Testing notes
Test steps are in [the
ticket](https://researchspace.atlassian.net/browse/RSDEV-296)

---------

Co-authored-by: Robert Lamacraft <robert.lamacraft@researchspace.com>