Rsdev 661 fix slow tests (#554)
## Description ##
This PR removes and updates some of the slowest unit and spring
transactional tests when running `mvn clean test`.

There's a package `testsandbox` which contains some exploratory code
around different libraries, some of which was ignored and most which
didn't test any rspace code and/or make any assertions. One of the tests
took over 10s to run (`DelayQueueTest`). I removed the package.

I also refactored some email code so that the retry interval was
configurable, so that `EmailBroadcastTest` can set to a low interval. In
this case, we use 10ms rather than 1s and save ~6s when running the
test. I also tidied up the code in this class, removing unused params
etc.

In `EmailSenderTest` we override the retry behaviour, since that's not
tested in any way in that class, and save ~3s.

Altogether, we save 20s which is over 5% for me running locally, where
`mvn clean test` previously took around 6mins.