Rsdev 907 link stoich to inventory (#581)
https://researchspace.atlassian.net/browse/RSDEV-907

## Description ##
This PR adds the backend CRUD endpoints for managing stoichiometry
inventory links.

These links are also included in the response from the `/stoichiometry`
endpoints.

Each change (create, update or delete) of a link generates a new
revision of the parent `Stoichiometry` entity, which can be retrieved
via the `/stoichiometry` endpoints.

Copied documents retain the link information, creating a new link to the
same inventory item with the same quantity as was in the original.

## Notes
1. Document copy for Stoichiometry tables always takes the latest
Stoichiometry version. This is an oversight/bug which will be handled
separately.
2. Links don't reduce any amounts in inventory or check there's enough
of an inventory compound. This will be handled via
https://researchspace.atlassian.net/browse/RSDEV-876

## Design decisions
The GET `/stoichiometry/link` endpoint doesn't take a revision - it
always returns the latest. Revisions are handled via the parent
`Stoichiometry` entity, since there probably won't be any need for users
or the front-end to retrieve stoichiometry link revisions outwith the
main `Stoichiometry` entity.

## Testing notes
Test scenarios are in the [jira
ticket](https://researchspace.atlassian.net/browse/RSDEV-907)

rspace-core-model changes:
https://github.com/rspace-os/rspace-core-model/pull/48