Rsdev 938 fix reaction doc import (#596)
https://researchspace.atlassian.net/browse/RSDEV-938

## Description ##
There's concept of "default" format for chemistry, which was used in the
previous closed-source chemistry and kept for compatibility while we
switched to open-source. However, we don't store all chemicals in that
format any longer in the database. Instead we (mostly) keep chemistry
files in their original formats.

In the import code, when importing a document with chemistry, we were
still attempting to convert to the default format. This would fail when
converting Ketcher reaction drawings.

The fix is to not try to convert to default format. We know the chemical
has previously (in the original, exported doc) been imported fine into
rspace, so we can import it as-is without applying any conversions.

## Testing notes
See steps in ticket.